Don't use perlbrew in a perlbrew lib
Fixes GH #6

The generated test files scrub environment variables like PERL5LIB, so
if one were to use a perlbrew script installed in a local lib, it
would look for App::perlbrew in that same local lib.  However, since
we have scrubbed PERL5LIB by that point, it would fail to find it.
This patch scrubs any local lib paths from PATH so that only a perlbrew
script outside of a perlbrew lib can be used.  There should be at least
one, considering there should be a perlbrew to create the perlbrew
lib!